// src/app/api/api-hotels/models/rq-property-delete-ce-request-1.ts
/* tslint:disable */
/* eslint-disable */
import { CeRequestCaller } from './ce-request-caller';
import { RqPropertyDelete } from './rq-property-delete';
export interface RqPropertyDeleteCeRequest1 {
  Caller: CeRequestCaller;
  MuteExceptions: boolean;
  Parameter: RqPropertyDelete;
  TrackingCode: string;
}